+++ apps/playground-nuxt/layouts/default.vue
@@ -2,14 +2,10 @@
 
 <template>
   <div>
-    <TheHeader />
-    <NuxtLoadingIndicator />
-    <main
-      class="bg-(--ui-bg) min-h-[calc(100vh-var(--ui-header-height))]"
-    >
+    <UMain>
       <slot />
-    </main>
+    </UMain>
   </div>
 </template>
 
-<style scoped></style>ß
+<style scoped></style>